+++ modules/src/signal/splot.c
@@ -69,8 +69,8 @@ static void getrule(double a, double b, double *ofs, double *step, int *nsub)
 /* compute max interval [n1,n2] such that [sx1,sx2] contains x([n1,n2]) */
 static void interval(int *n1, int *n2)
 {
-  *n1 = (int)ceil ((sx1-(double)signal->shift)/(double)signal->scale);
-  *n2 = (int)floor((sx2-(double)signal->shift)/(double)signal->scale);
+  *n1 = ceil ((sx1-(double)signal->shift)/(double)signal->scale);
+  *n2 = floor((sx2-(double)signal->shift)/(double)signal->scale);
   if (*n1<0) *n1=0;
   if (*n2>=signal->size) *n2=signal->size-1;
 }
@@ -122,10 +122,10 @@ static void draw_framed_ccimage(Ccimage u, int xa, int ya, int xb, int yb, unsig
     if (tya>txa) txa=tya;
     if (tyb<txb) txb=tyb;
     if (txa<txb) {
-      xb = xa + (int) floor(txb * dx + .5);
-      yb = ya + (int) floor(txb * dy + .5);
-      xa +=     (int) floor(txa * dx + .5);
-      ya +=     (int) floor(txa * dy + .5);
+      xb = xa + floor(txb * dx + .5);
+      yb = ya + floor(txb * dy + .5);
+      xa +=     floor(txa * dx + .5);
+      ya +=     floor(txa * dy + .5);
       if (xa>=xmin && xa<=xmax && ya>=ymin && ya<=ymax && 
 	  xb>=xmin && xb<=xmax && yb>=ymin && yb<=ymax) 
 	mw_draw_ccimage(u,xa,ya,xb,yb,r,g,b);
